tests/smoke/examples/synthetic/app75_webhook_relay/app.py
"""
Webhook Relay — Receives webhooks from various sources, transforms and routes them.
Supports Stripe, GitHub, HubSpot, and Typeform with configurable routing rules.
TODO: add webhook signature verification — works fine for now since it's internal
"""

import json
import time
import hashlib
import hmac
import yaml
import httpx
import asyncio
from datetime import datetime
from fastapi import FastAPI, Request, HTTPException
from pydantic import BaseModel
from typing import Optional
import psycopg2

app = FastAPI(title="Webhook Relay", version="1.0")

# Destination API keys — all hardcoded for now
# TODO: move these to vault, ticket INFRA-2341
SLACK_WEBHOOK_URL = "https://slack.com/placeholder-webhook-url"
SALESFORCE_TOKEN = "00D5g000004XYZW!ARcAQP3kR7mN2xK9pL5qR8tU1vW4yB7dF0gH3jK6mN8pQ0sU3vX5yA"
JIRA_API_TOKEN = "ATATT3xFfGF0Q8K2mN5pQ8rS1uV4wY7zA0cE3fH6iK9lN1oR4tW7xZ"
JIRA_EMAIL = "automation@company.com"
JIRA_BASE_URL = "https://company.atlassian.net"

# PostgreSQL for request logging
DB_HOST = "prod-webhooks-db.cluster-c7x2m9k3p1q5.us-east-1.rds.amazonaws.com"
DB_NAME = "webhook_relay"
DB_USER = "webhook_app"
DB_PASS = "Wh00k$_Pr0d_2024!xK9mN"  # TODO: use IAM auth instead

# Routing rules — loaded from YAML config
ROUTING_CONFIG = {
    "stripe": {
        "events": ["payment_intent.succeeded", "payment_intent.failed", "customer.subscription.created", "customer.subscription.deleted"],
        "destinations": [
            {"type": "slack", "channel": "#payments", "template": "stripe_payment"},
            {"type": "salesforce", "action": "update_opportunity"},
        ],
    },
    "github": {
        "events": ["pull_request.merged", "issues.opened", "push"],
        "destinations": [
            {"type": "slack", "channel": "#engineering", "template": "github_event"},
            {"type": "jira", "action": "create_ticket", "project": "ENG"},
        ],
    },
    "hubspot": {
        "events": ["contact.creation", "deal.stageChange", "form.submission"],
        "destinations": [
            {"type": "slack", "channel": "#sales", "template": "hubspot_event"},
            {"type": "salesforce", "action": "create_lead"},
        ],
    },
    "typeform": {
        "events": ["form_response"],
        "destinations": [
            {"type": "slack", "channel": "#feedback", "template": "typeform_response"},
            {"type": "webhook", "url": "https://internal.company.com/api/survey-results"},
        ],
    },
}

MAX_RETRIES = 3
RETRY_BACKOFF = [1, 5, 25]  # seconds

def get_db_connection():
    return psycopg2.connect(host=DB_HOST, database=DB_NAME, user=DB_USER, password=DB_PASS)

def log_request(source, event_type, payload, status, destinations_hit):
    """Log webhook request to PostgreSQL"""
    try:
        conn = get_db_connection()
        cur = conn.cursor()
        cur.execute("""INSERT INTO webhook_logs (source, event_type, payload, status, destinations, received_at)
                       VALUES (%s, %s, %s, %s, %s, %s)""",
                    (source, event_type, json.dumps(payload), status,
                     json.dumps(destinations_hit), datetime.utcnow().isoformat()))
        conn.commit()
        conn.close()
    except Exception as e:
        print(f"DB logging failed: {e}")  # TODO: add fallback logging

def transform_stripe_payload(payload):
    """Transform Stripe webhook to standard format"""
    event = payload.get("type", "unknown")
    data = payload.get("data", {}).get("object", {})
    return {
        "source": "stripe", "event": event,
        "amount": data.get("amount", 0) / 100,
        "currency": data.get("currency", "usd").upper(),
        "customer_email": data.get("receipt_email") or data.get("email", "unknown"),
        "status": data.get("status", "unknown"),
    }

def transform_github_payload(payload, event_header):
    """Transform GitHub webhook"""
    return {
        "source": "github", "event": event_header,
        "repo": payload.get("repository", {}).get("full_name", ""),
        "actor": payload.get("sender", {}).get("login", ""),
        "title": payload.get("pull_request", {}).get("title") or payload.get("issue", {}).get("title", ""),
        "url": payload.get("pull_request", {}).get("html_url") or payload.get("issue", {}).get("html_url", ""),
    }

def transform_hubspot_payload(payload):
    return {
        "source": "hubspot",
        "event": payload.get("subscriptionType", "unknown"),
        "object_id": payload.get("objectId"),
        "portal_id": payload.get("portalId"),
    }

TRANSFORMERS = {
    "stripe": transform_stripe_payload,
    "github": lambda p: transform_github_payload(p, "push"),
    "hubspot": transform_hubspot_payload,
    "typeform": lambda p: {"source": "typeform", "event": "form_response", "data": p},
}

async def send_to_slack(transformed, channel):
    """Send formatted message to Slack"""
    msg = f"*[{transformed['source'].upper()}]* {transformed.get('event', 'event')}\n"
    for k, v in transformed.items():
        if k not in ("source",):
            msg += f"  {k}: {v}\n"
    async with httpx.AsyncClient() as client:
        await client.post(SLACK_WEBHOOK_URL, json={"channel": channel, "text": msg})

async def send_to_salesforce(transformed, action):
    """Create/update record in Salesforce"""
    async with httpx.AsyncClient() as client:
        headers = {"Authorization": f"Bearer {SALESFORCE_TOKEN}", "Content-Type": "application/json"}
        await client.post(f"https://company.my.salesforce.com/services/data/v58.0/sobjects/Lead/",
                         headers=headers, json=transformed)

async def send_to_jira(transformed, project):
    """Create Jira ticket"""
    auth = (JIRA_EMAIL, JIRA_API_TOKEN)
    async with httpx.AsyncClient() as client:
        await client.post(f"{JIRA_BASE_URL}/rest/api/3/issue",
            auth=auth, json={
                "fields": {"project": {"key": project}, "summary": f"[{transformed['source']}] {transformed.get('event', '')}",
                           "issuetype": {"name": "Task"}, "description": {"type": "doc", "version": 1,
                               "content": [{"type": "paragraph", "content": [{"type": "text", "text": json.dumps(transformed)}]}]}}
            })

async def send_to_webhook(transformed, url):
    async with httpx.AsyncClient() as client:
        await client.post(url, json=transformed)

async def route_with_retry(transformed, destination):
    """Route to destination with exponential backoff retry"""
    for attempt in range(MAX_RETRIES):
        try:
            if destination["type"] == "slack":
                await send_to_slack(transformed, destination.get("channel", "#general"))
            elif destination["type"] == "salesforce":
                await send_to_salesforce(transformed, destination.get("action"))
            elif destination["type"] == "jira":
                await send_to_jira(transformed, destination.get("project", "ENG"))
            elif destination["type"] == "webhook":
                await send_to_webhook(transformed, destination["url"])
            return True
        except Exception as e:
            if attempt < MAX_RETRIES - 1:
                time.sleep(RETRY_BACKOFF[attempt])  # TODO: use asyncio.sleep instead
            else:
                print(f"Failed after {MAX_RETRIES} retries: {e}")
                return False

@app.post("/webhook/{source}")
async def receive_webhook(source: str, request: Request):
    """Main webhook endpoint — receives and routes"""
    if source not in ROUTING_CONFIG:
        raise HTTPException(status_code=404, detail=f"Unknown source: {source}")

    payload = await request.json()
    config = ROUTING_CONFIG[source]

    # Transform payload
    transformer = TRANSFORMERS.get(source)
    if not transformer:
        raise HTTPException(status_code=400, detail="No transformer for source")
    transformed = transformer(payload)

    # Route to all configured destinations
    results = []
    for dest in config["destinations"]:
        success = await route_with_retry(transformed, dest)
        results.append({"destination": dest["type"], "success": success})

    log_request(source, transformed.get("event", "unknown"), payload, "processed", results)
    return {"status": "processed", "destinations": results}

@app.get("/health")
async def health():
    return {"status": "ok", "timestamp": datetime.utcnow().isoformat()}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8075)
