tests/smoke/examples/synthetic/app75_webhook_relay/requirements.txt
fastapi==0.109.0
uvicorn[standard]
httpx
psycopg2-binary
pyyaml
pydantic
